Refactored Random Event Unit Tests to Match New Package Structure
