+++ MekHQ/unittests/mekhq/campaign/randomEvents/personalities/AggressionTest.java
@@ -30,8 +30,10 @@
  * <https://www.xbox.com/en-US/developers/rules> and it is not endorsed by or
  * affiliated with Microsoft.
  */
-package mekhq.campaign.randomEvents.personalities.enums;
+package mekhq.campaign.randomEvents.personalities;
 
+import static mekhq.campaign.randomEvents.personalities.Aggression.MAXIMUM_VARIATIONS;
+import static mekhq.campaign.randomEvents.personalities.Aggression.NONE;
 import static mekhq.utilities.MHQInternationalization.isResourceKeyValid;
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertTrue;
@@ -41,47 +43,45 @@ import java.util.stream.IntStream;
 import java.util.stream.Stream;
 
 import megamek.common.enums.Gender;
-import mekhq.campaign.randomEvents.personalities.Greed;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.Arguments;
 import org.junit.jupiter.params.provider.CsvSource;
 import org.junit.jupiter.params.provider.EnumSource;
 import org.junit.jupiter.params.provider.MethodSource;
 
-public class GreedTest {
+public class AggressionTest {
     @ParameterizedTest
-    @CsvSource(value = { "ASTUTE,ASTUTE", "INVALID_STATUS,NONE", "'',NONE", "'null',NONE", "1,ASTUTE" })
-    void testFromStringVariousInputs(String input, Greed expected) {
+    @CsvSource(value = { "AGGRESSIVE,AGGRESSIVE", "INVALID_STATUS,NONE", "'',NONE", "'null',NONE", "1,AGGRESSIVE" })
+    void testFromStringVariousInputs(String input, Aggression expected) {
         if ("null".equals(input)) {
             input = null;
         }
-        Greed result = Greed.fromString(input);
+        Aggression result = Aggression.fromString(input);
         assertEquals(expected, result);
     }
 
     @ParameterizedTest
-    @EnumSource(value = Greed.class)
-    void testFromString_Ordinal_All(Greed value) {
-        Greed result = Greed.fromString(String.valueOf(value.ordinal()));
+    @EnumSource(value = Aggression.class)
+    void testFromString_Ordinal_All(Aggression value) {
+        Aggression result = Aggression.fromString(String.valueOf(value.ordinal()));
         assertEquals(value, result);
     }
 
     @ParameterizedTest
-    @EnumSource(value = Greed.class)
-    void testGetLabel_notInvalid(Greed status) {
+    @EnumSource(value = Aggression.class)
+    void testGetLabel_notInvalid(Aggression status) {
         String label = status.getLabel();
         assertTrue(isResourceKeyValid(label));
     }
 
-    static Stream<Arguments> provideGreedsAndIndices() {
-        return Arrays.stream(Greed.values())
-                     .flatMap(trait -> IntStream.range(0, Greed.MAXIMUM_VARIATIONS)
-                                             .mapToObj(i -> Arguments.of(trait, i)));
+    static Stream<Arguments> provideAggressionsAndIndices() {
+        return Arrays.stream(Aggression.values())
+                     .flatMap(trait -> IntStream.range(0, MAXIMUM_VARIATIONS).mapToObj(i -> Arguments.of(trait, i)));
     }
 
     @ParameterizedTest
-    @MethodSource(value = "provideGreedsAndIndices")
-    void testGetDescription_notInvalid(Greed trait, int i) {
+    @MethodSource(value = "provideAggressionsAndIndices")
+    void testGetDescription_notInvalid(Aggression trait, int i) {
         String description = trait.getDescription(i, Gender.MALE, "Barry");
         assertTrue(isResourceKeyValid(description));
     }
@@ -89,13 +89,13 @@ public class GreedTest {
     @ParameterizedTest
     @CsvSource(value = { "99", "1000", "-1" })
     void testGetDescription_InvalidDescriptionIndex(int invalidIndex) {
-        String description = Greed.NONE.getDescription(invalidIndex, Gender.MALE, "Barry");
+        String description = NONE.getDescription(invalidIndex, Gender.MALE, "Barry");
         assertTrue(isResourceKeyValid(description));
     }
 
     @ParameterizedTest
-    @EnumSource(value = Greed.class)
-    void testGetRoninMessage_notInvalid(Greed trait) {
+    @EnumSource(value = Aggression.class)
+    void testGetRoninMessage_notInvalid(Aggression trait) {
         String description = trait.getRoninMessage("Commander");
         assertTrue(isResourceKeyValid(description));
     }
