+++ MekHQ/unittests/mekhq/campaign/randomEvents/prisoners/MobTypeTest.java
@@ -1,5 +1,5 @@
 /*
- * Copyright (C) 2025 The MegaMek Team. All Rights Reserved.
+ * Copyright (C) 2025-2026 The MegaMek Team. All Rights Reserved.
  *
  * This file is part of MekHQ.
  *
@@ -30,20 +30,33 @@
  * <https://www.xbox.com/en-US/developers/rules> and it is not endorsed by or
  * affiliated with Microsoft.
  */
-package mekhq.campaign.randomEvents.personalities.enums;
+package mekhq.campaign.randomEvents.prisoners;
 
-import static mekhq.utilities.MHQInternationalization.isResourceKeyValid;
-import static org.junit.jupiter.api.Assertions.assertTrue;
+import static org.junit.jupiter.api.Assertions.assertEquals;
 
-import mekhq.campaign.randomEvents.personalities.PersonalityTraitType;
-import org.junit.jupiter.params.ParameterizedTest;
-import org.junit.jupiter.params.provider.EnumSource;
+import mekhq.campaign.randomEvents.prisoners.prisonerEvents.MobType;
+import org.junit.jupiter.api.Test;
 
-class PersonalityTraitTypeTest {
-    @ParameterizedTest
-    @EnumSource(value = PersonalityTraitType.class)
-    void testGetLabel_notInvalid(PersonalityTraitType status) {
-        String label = status.getLabel();
-        assertTrue(isResourceKeyValid(label));
+/**
+ * A test class for validating the behavior and properties of the {@code MobType} enumeration.
+ *
+ * <p>This class contains unit tests to ensure that the ranges defined by each {@code MobType}
+ * are consistent and correctly implemented. It checks whether the minimum value of each subsequent {@code MobType} is
+ * one greater than the maximum value of the previous {@code MobType}.</p>
+ *
+ * <p>These tests aim to validate that the {@code MobType} enumerations are properly sequential and
+ * follow the expected logical configuration.</p>
+ */
+class MobTypeTest {
+    @Test
+    void testToStringSmall() {
+        int maximum = 0;
+        for (MobType mobType : MobType.values()) {
+            int minimum = mobType.getMinimum();
+
+            assertEquals(maximum + 1, minimum);
+
+            maximum = mobType.getMaximum();
+        }
     }
 }
