+++ MekHQ/unittests/mekhq/campaign/randomEvents/prisoners/PrisonerCaptureStyleTest.java
@@ -1,5 +1,5 @@
 /*
- * Copyright (C) 2025-2026 The MegaMek Team. All Rights Reserved.
+ * Copyright (C) 2025 The MegaMek Team. All Rights Reserved.
  *
  * This file is part of MekHQ.
  *
@@ -30,33 +30,34 @@
  * <https://www.xbox.com/en-US/developers/rules> and it is not endorsed by or
  * affiliated with Microsoft.
  */
-package mekhq.campaign.randomEvents.prisoners.enums;
+package mekhq.campaign.randomEvents.prisoners;
 
-import static org.junit.jupiter.api.Assertions.assertEquals;
+import static mekhq.utilities.MHQInternationalization.isResourceKeyValid;
+import static org.junit.jupiter.api.Assertions.assertTrue;
 
-import mekhq.campaign.randomEvents.prisoners.prisonerEvents.MobType;
 import org.junit.jupiter.api.Test;
 
 /**
- * A test class for validating the behavior and properties of the {@code MobType} enumeration.
+ * A test class for validating the functionality of the {@code PrisonerCaptureStyle} enumeration.
  *
- * <p>This class contains unit tests to ensure that the ranges defined by each {@code MobType}
- * are consistent and correctly implemented. It checks whether the minimum value of each subsequent {@code MobType} is
- * one greater than the maximum value of the previous {@code MobType}.</p>
- *
- * <p>These tests aim to validate that the {@code MobType} enumerations are properly sequential and
- * follow the expected logical configuration.</p>
+ * <p>This class contains unit tests to ensure that each {@code PrisonerCaptureStyle} has valid
+ * resource keys for its labels and tooltips. The tests verify that no invalid labels or tooltips are present, as
+ * determined by the {@code isResourceKeyValid} method.</p>
  */
-class MobTypeTest {
+public class PrisonerCaptureStyleTest {
     @Test
-    void testToStringSmall() {
-        int maximum = 0;
-        for (MobType mobType : MobType.values()) {
-            int minimum = mobType.getMinimum();
+    public void testGetLabel_notInvalid() {
+        for (PrisonerCaptureStyle status : PrisonerCaptureStyle.values()) {
+            String label = status.getLabel();
+            assertTrue(isResourceKeyValid(label));
+        }
+    }
 
-            assertEquals(maximum + 1, minimum);
-
-            maximum = mobType.getMaximum();
+    @Test
+    public void testGetTitleExtension_notInvalid() {
+        for (PrisonerCaptureStyle status : PrisonerCaptureStyle.values()) {
+            String titleExtension = status.getTooltip();
+            assertTrue(isResourceKeyValid(titleExtension));
         }
     }
 }
